requirements.txt
torch==2.0.0
torchvision==0.15.0
torchaudio==0.15.0
numpy>=1.23.0
Pillow>=9.0.0
opencv-python>=4.6.0
scipy>=1.9.0
tqdm>=4.64.0
matplotlib>=3.5.0
scikit-image>=0.19.0
imageio>=2.9.0
scikit-learn>=1.2.0
psutil>=5.9.0
GPUtil>=1.4.0
